refactor: ♻️  rely on jsonnet binary instead of go package
for now, working with the jsonnet go API feels premature. To get the most out of Jsonnet and not rewrite code to manage JSONNET_PATH and other things, we'll just shell out to the jsonnet external binary. As long as this is wrapped in behind a abstracted interface, I think this will be easy to change later.